Support tcp network between runner server and runner

+++ cmd/bb_runner/main.go
@@ -49,13 +49,16 @@ func main() {
 	s := grpc.NewServer()
 	runner.RegisterRunnerServer(s, runnerServer)
 
-	if err := os.Remove(runnerConfiguration.ListenPath); err != nil && !os.IsNotExist(err) {
-		log.Fatalf("Could not remove stale socket %#v: %s", runnerConfiguration.ListenPath, err)
+	if runnerConfiguration.ListenNetwork == "unix" {
+		if err := os.Remove(runnerConfiguration.ListenPath); err != nil && !os.IsNotExist(err) {
+			log.Fatalf("Could not remove stale socket %#v: %s", runnerConfiguration.ListenPath, err)
+		}
 	}
 
-	sock, err := net.Listen("unix", runnerConfiguration.ListenPath)
+	sock, err := net.Listen(runnerConfiguration.ListenNetwork, runnerConfiguration.ListenPath)
 	if err != nil {
-		log.Fatalf("Failed to create listening socket %#v: %s", runnerConfiguration.ListenPath, err)
+		log.Fatalf("Failed to create %#v listening socket %#v: %s",
+			runnerConfiguration.ListenNetwork, runnerConfiguration.ListenPath, err)
 	}
 	if err := s.Serve(sock); err != nil {
 		log.Fatal("Failed to serve RPC server: ", err)
